Fixed for loops to get rid of warnings, deleted unused code in initial placement, and made macros also not use free_locations and legal_pos

+++ vpr/src/place/compressed_grid.cpp
@@ -155,7 +155,7 @@ void echo_compressed_grids(char* filename, const std::vector<t_compressed_block_
     fprintf(fp, "--------------------------------------------------------------\n");
     fprintf(fp, "\n");
 
-    for (int i = 0; i < comp_grids.size(); i++) {
+    for (int i = 0; i < (int)comp_grids.size(); i++) {
     	fprintf(fp, "\n\nGrid type: %s \n", device_ctx.logical_block_types[i].name);
 
     	fprintf(fp, "X coordinates: \n");
